pyproject.toml
[tool.poetry]
name = "VKBot"
version = "2.1.0"
description = ""
authors = ["Jônathas Gouveia <jonathas_gouv@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pycrypto = "^2.6.1"
lxml = "^4.6.2"
unidecode = "^1.1.2"
bs4 = "^0.0.1"
firebase = "^3.0.1"
flask = "^1.1.2"
python-jwt = "^3.3.0"
gcloud = "^0.18.3"
sseclient = "^0.0.27"
youtube-dl = "^2020.7"
requests-toolbelt = "^0.9.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
